feat(ai-auditing): introduce paid AI auditing workflow, update entities, services, jobs, events, and tests Add AI auditing paid service and integrate into competition flow Extend Competition with AI auditing attributes (e.g., ai_auditing, auditing_time_for_level) Update Level model, service, and controller to support AI auditing and reassignment flows Enhance finish-level process to trigger AI auditing pipeline Create jobs to run AI auditing tasks and background processing Fire events for balance deduction and auditing lifecycle Add repository methods to reassign users’ responses to level creator and auto-confirm AI-generated responses

+++ app/Models/Payment/CoinBalance.php
@@ -6,6 +6,35 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\MorphTo;
 
+/**
+ * 
+ *
+ * @property int $id
+ * @property int $balanceable_id
+ * @property string $balanceable_type
+ * @property int $balance
+ * @property int $total_earned
+ * @property int $total_spent
+ * @property \Illuminate\Support\Carbon|null $created_at
+ * @property \Illuminate\Support\Carbon|null $updated_at
+ * @property-read Model|\Eloquent $balanceable
+ * @property-read string $balance_formatted
+ * @property-read string $total_earned_formatted
+ * @property-read string $total_spent_formatted
+ * @method static \Database\Factories\Payment\CoinBalanceFactory factory($count = null, $state = [])
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinBalance newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinBalance newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinBalance query()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinBalance whereBalance($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinBalance whereBalanceableId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinBalance whereBalanceableType($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinBalance whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinBalance whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinBalance whereTotalEarned($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinBalance whereTotalSpent($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinBalance whereUpdatedAt($value)
+ * @mixin \Eloquent
+ */
 class CoinBalance extends Model
 {
     use HasFactory;
@@ -18,9 +47,9 @@ class CoinBalance extends Model
     ];
 
     protected $casts = [
-        'balance' => 'integer',
-        'total_earned' => 'integer',
-        'total_spent' => 'integer',
+        'balance' => 'float',
+        'total_earned' => 'float',
+        'total_spent' => 'float',
         'created_at' => 'datetime',
         'updated_at' => 'datetime'
     ];
