+++ app/Models/Payment/CoinPricing.php
@@ -9,6 +9,47 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasOne;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 
+/**
+ * 
+ *
+ * @property int $id
+ * @property string $name Short, unique name for pricing rule
+ * @property string $display_name Optional longer description
+ * @property string $user_type Type of user this pricing applies to
+ * @property numeric $base_amount Money amount in DZD
+ * @property int $base_coins Number of coins given for this amount
+ * @property bool $is_active Whether this pricing is currently active
+ * @property int|null $created_by_admin_id
+ * @property \Illuminate\Support\Carbon|null $created_at
+ * @property \Illuminate\Support\Carbon|null $updated_at
+ * @property-read \App\Models\Payment\CoinOffer|null $activeOffer
+ * @property-read Admin|null $createdByAdmin
+ * @property-read float $coins_per_dzd
+ * @property-read string $formatted_amount
+ * @property-read string $formatted_coins
+ * @property-read string $full_name
+ * @property-read string $rate_description
+ * @property-read string $user_type_label
+ * @property-read \Illuminate\Database\Eloquent\Collection<int, \App\Models\Payment\CoinOffer> $offers
+ * @property-read int|null $offers_count
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing active()
+ * @method static \Database\Factories\Payment\CoinPricingFactory factory($count = null, $state = [])
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing forUserType(string $userType)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing query()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing whereBaseAmount($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing whereBaseCoins($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing whereCreatedByAdminId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing whereDisplayName($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing whereIsActive($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing whereName($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing whereUpdatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|CoinPricing whereUserType($value)
+ * @mixin \Eloquent
+ */
 class CoinPricing extends Model
 {
     use HasFactory;
@@ -25,7 +66,7 @@ class CoinPricing extends Model
 
     protected $casts = [
         'base_amount' => 'decimal:2',
-        'base_coins' => 'integer',
+        'base_coins' => 'float',
         'is_active' => 'boolean',
         'created_at' => 'datetime',
         'updated_at' => 'datetime'
